requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
dj-database-url==3.0.1
Django==4.2.30
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
packaging==26.1
psycopg2-binary==2.9.11
requests==2.32.5
resend==2.28.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
whitenoise==6.11.0
Pillow>=10.0,<12
